- adding Create Item  - adding Read Item

+++ app/src/main/java/com/artevak/kasirpos/data/model/Keranjang.kt
@@ -6,12 +6,12 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class Keranjang(
     var nama_barang: String? = "",
-    var jumlah_beli: Int? = 0,
-    var harga_barang: Int? = 0,
+    var jumlah_beli: Long? = 0,
+    var harga_barang: Long? = 0,
     var deskripsi: String? = "",
     var picture: String? = "",
     var satuan: String? = "",
-    var subtotal: Int? = 0,
+    var subtotal: Long? = 0,
     var id_barang: String? = "",
-    var untung: Int? = 0,
+    var untung: Long? = 0,
 ) : Parcelable
